milestone

+++ Raytrace.cpp
@@ -34,7 +34,7 @@ glm::vec3 Raytrace::calculateRay(vec3& eye, vec3& center, vec3& up, float fovx, 
   vec3 w = glm::normalize(eye - center);
   vec3 u = glm::normalize(glm::cross(up, w));
   vec3 v = glm::cross(w, u);
-  float a = glm::tan(glm::radians(fovx/2)) * ((j-(width/2.0))/(width/2.0));
-  float b = glm::tan(glm::radians(fovy/2)) * (((height/2.0)-i)/(height/2.0));
+  float a = glm::tan(glm::radians(fovx/2)) * ((j-(static_cast<float>(width)/2))/(static_cast<float>(width)/2));
+  float b = glm::tan(glm::radians(fovy/2)) * (((static_cast<float>(height)/2)-i)/(static_cast<float>(height)/2));
   return glm::normalize(a*u + b*v - w);
 }
